+++ Sources/Views/Settings/TTS/TTSDictionaryEditView.swift
@@ -249,6 +249,7 @@ struct TTSDictionaryEditView: View {
                 }
             }
         }
+        .tint(.white)
     }
 
     private func loadDictionary() async {