+++ Sources/Views/TTSWidget/TTSSettingsSheet.swift
@@ -10,5 +10,6 @@ struct TTSSettingsSheet: View {
         NavigationStack {
             TTSSettingsView(isPresentedAsSheet: true)
         }
+        .tint(.white)
     }
 }